Refactor: 서비스 메서드 분리로 인한 코드 리팩토링
