+++ src/main/java/com/dangdangsalon/domain/estimate/request/service/GroomerEstimateRequestService.java
@@ -25,7 +25,6 @@ import java.util.List;
 import java.util.Optional;
 
 // groomer 관련 서비스
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class GroomerEstimateRequestService {
@@ -33,8 +32,7 @@ public class GroomerEstimateRequestService {
     private final GroomerServiceAreaRepository groomerServiceAreaRepository;
     private final GroomerCanServiceRepository groomerCanServiceRepository;
     private final GroomerEstimateRequestRepository groomerEstimateRequestRepository;
-    private final NotificationService notificationService;
-    private final UserRepository userRepository;
+    private final GroomerEstimateRequestNotificationService groomerEstimateRequestNotificationService;
 
     @Transactional
     public void insertGroomerEstimateRequests(EstimateRequest estimateRequest, District district, EstimateRequestDto estimateRequestDto) {
@@ -121,32 +119,6 @@ public class GroomerEstimateRequestService {
                 .build();
 
         groomerEstimateRequestRepository.save(groomerEstimateRequest);
-        sendNotificationToGroomer(estimateRequest, groomerProfile);
-    }
-
-    private void sendNotificationToGroomer(EstimateRequest estimateRequest, GroomerProfile groomerProfile) {
-
-        Long userId = groomerProfile.getUser().getId();
-
-        User user = userRepository.findById(userId).orElseThrow(() ->
-                new IllegalArgumentException("사용자를 찾을 수 없습니다: " + userId)
-        );
-
-        if (Boolean.FALSE.equals(user.getNotificationEnabled())) {
-            log.info("알림 비활성화: " + user.getId());
-        }else{
-            Optional<String> optionalFcmToken = notificationService.getFcmToken(userId);
-
-            if (optionalFcmToken.isPresent()) {
-                String fcmToken = optionalFcmToken.get();
-
-                String title = "새로운 견적 요청";
-                String body = "새로운 견적 요청이 도착했습니다. 확인하세요.";
-                // 알림 전송
-                notificationService.sendNotificationWithData(fcmToken, title, body, "견적 요청", estimateRequest.getId());
-                // redis 에 알림 내용 저장
-                notificationService.saveNotificationToRedis(userId, title, body, "견적 요청", estimateRequest.getId());
-            }
-        }
+        groomerEstimateRequestNotificationService.sendNotificationToGroomer(estimateRequest, groomerProfile);
     }
 }
